🌐 i18n: AgentProfileList/DataVendors 内联 UI 改为 t() 调用
- AgentProfileList: 收起/高级/任务指令/温度/最大工具轮数 → t()
- DataVendorsTab: 9 个状态/提示/描述标签 → t()
- post-merge-stock: 新增 22 个 profile/vendors 子键 (zh-CN + en-US)

+++ src-tauri/crates/astock-data/src/vendors/eastmoney.rs
@@ -631,7 +631,7 @@ impl StockVendor for EastMoneyVendor {
                         .and_then(|v| v.as_i64())
                         .map(|ts| {
                             let secs = ts / 1000;
-                            let naive = chrono::DateTime::from_timestamp(secs as i64, 0)
+                            let naive = chrono::DateTime::from_timestamp(secs, 0)
                                 .map(|dt| dt.format("%Y-%m-%d").to_string())
                                 .unwrap_or_default();
                             naive
